+++ interop/kotlin-metadata/src/test/kotlin/com/squareup/kotlinpoet/metadata/specs/KotlinPoetMetadataSpecsTest.kt
@@ -1750,7 +1750,7 @@ class KotlinPoetMetadataSpecsTest : MultiClassInspectorTest() {
         """
         @kotlin.SinceKotlin(version = "1.3")
         @kotlin.`annotation`.Retention(value = kotlin.`annotation`.AnnotationRetention.RUNTIME)
-        @kotlin.`annotation`.Target(allowedTargets = arrayOf(kotlin.`annotation`.AnnotationTarget.CLASS))
+        @kotlin.`annotation`.Target(allowedTargets = [kotlin.`annotation`.AnnotationTarget.CLASS])
         public annotation class Metadata(
           @get:kotlin.jvm.JvmName(name = "k")
           public val kind: kotlin.Int = throw NotImplementedError("Stub!"),
@@ -1787,7 +1787,7 @@ class KotlinPoetMetadataSpecsTest : MultiClassInspectorTest() {
       .isEqualTo(
         """
         @kotlin.`annotation`.Retention(value = kotlin.`annotation`.AnnotationRetention.RUNTIME)
-        @kotlin.`annotation`.Target(allowedTargets = arrayOf(kotlin.`annotation`.AnnotationTarget.CLASS))
+        @kotlin.`annotation`.Target(allowedTargets = [kotlin.`annotation`.AnnotationTarget.CLASS])
         public annotation class Metadata(
           @get:kotlin.jvm.JvmName(name = "k")
           public val kind: kotlin.Int = throw NotImplementedError("Stub!"),